+++ backend/app/models/exam_record.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, BigInteger, DECIMAL, JSON, DateTime, ForeignKey, UniqueConstraint
+from sqlalchemy import Column, Integer, DECIMAL, JSON, DateTime, ForeignKey, UniqueConstraint
 from sqlalchemy.orm import relationship
 
 from app.database import Base
@@ -10,9 +10,9 @@ class ExamRecord(Base):
         UniqueConstraint("exam_id", "user_id", name="uk_exam_user"),
     )
 
-    id = Column(BigInteger, primary_key=True, autoincrement=True)
-    exam_id = Column(BigInteger, ForeignKey("exam.id", ondelete="CASCADE"), nullable=False)
-    user_id = Column(BigInteger, ForeignKey("sys_user.id", ondelete="CASCADE"), nullable=False, index=True)
+    id = Column(Integer, primary_key=True, autoincrement=True)
+    exam_id = Column(Integer, ForeignKey("exam.id", ondelete="CASCADE"), nullable=False)
+    user_id = Column(Integer, ForeignKey("sys_user.id", ondelete="CASCADE"), nullable=False, index=True)
     score = Column(DECIMAL(5, 1), default=None)
     answers = Column(JSON, default=None)
     start_time = Column(DateTime, nullable=False)
